// bamul/src/app/shared/model/VehicleDocuments .ts
export class VehicleDocuments {
    vehicleId: number;
    documentType: string;
    documentProviderName: string;
    documentNumber: string;
    docValidateFrom: any;
    docValidateTo: any;
    recordCreationDate: any;
    recordCreationBy: number;
    lastUpdateDate: any;
    lastUpdateBy: number;
}